+++ tests/unit/utils_testing_test.py
@@ -3,6 +3,7 @@ import math
 
 import numpy as np
 import pandas as pd
+import polars as pl
 import pytest
 from etoolbox.utils.testing import assert_equal, capture, idfn
 
@@ -12,6 +13,8 @@ from etoolbox.utils.testing import assert_equal, capture, idfn
     [
         (pd.Series([1, 2, 3]), pd.Series([1, 2, 3]), None),
         (pd.Series([1, 2, 3]), pd.Series([1, 45, 3]), AssertionError),
+        (pl.Series([1, 2, 3]), pl.Series([1, 2, 3]), None),
+        (pl.Series([1, 2, 3]), pl.Series([1, 45, 3]), AssertionError),
         ([1, pd.Series([1, 2, 3]), 3], [1, pd.Series([1, 2, 3]), 3], None),
         ([1, pd.Series([1, 2, 3]), 3], [1, pd.Series([1, 45, 3]), 3], AssertionError),
         ((1, 2, 3), (1, 2, 3), None),
@@ -47,6 +50,26 @@ from etoolbox.utils.testing import assert_equal, capture, idfn
                     {(0, "a"): [0.1, 2.1, 2.0], (0, "b"): [0.1, 3.1, 2.0]}
                 ),
             },
+            AssertionError,
+        ),
+        (
+            pl.DataFrame({"a": [0.1, 2.1, 2.0], "b": [0.1, 3.1, 2.0]}),
+            pl.DataFrame({"a": [0.1, 2.1, 2.0], "b": [0.1, 3.1, 2.0]}),
+            None,
+        ),
+        (
+            pl.DataFrame({"a": [0.1, 66.2, 2.0], "b": [0.1, 3.1, 2.0]}),
+            pl.DataFrame({"a": [0.1, 2.1, 2.0], "b": [0.1, 3.1, 2.0]}),
+            AssertionError,
+        ),
+        (
+            pl.LazyFrame({"a": [0.1, 2.1, 2.0], "b": [0.1, 3.1, 2.0]}),
+            pl.LazyFrame({"a": [0.1, 2.1, 2.0], "b": [0.1, 3.1, 2.0]}),
+            None,
+        ),
+        (
+            pl.LazyFrame({"a": [0.1, 66.2, 2.0], "b": [0.1, 3.1, 2.0]}),
+            pl.LazyFrame({"a": [0.1, 2.1, 2.0], "b": [0.1, 3.1, 2.0]}),
             AssertionError,
         ),
     ],